+++ buildkite/resource_pipeline.go
@@ -138,7 +138,6 @@ type pipelineResponse interface {
 	GetSlug() string
 	GetSteps() PipelineFieldsStepsPipelineSteps
 	GetTags() []PipelineFieldsTagsPipelineTag
-	GetWebhookURL() string
 }
 
 func newPipelineResource(archiveOnDelete *bool) func() resource.Resource {
@@ -216,6 +215,7 @@ func (p *pipelineResource) Create(ctx context.Context, req resource.CreateReques
 	log.Printf("Successfully created pipeline with id '%s'.", response.PipelineCreate.Pipeline.Id)
 
 	setPipelineModel(&state, &response.PipelineCreate.Pipeline)
+	state.WebhookUrl = types.StringValue(response.PipelineCreate.Pipeline.GetWebhookURL())
 
 	if plan.ProviderSettings != nil {
 		pipelineExtraInfo, err := updatePipelineExtraInfo(ctx, response.PipelineCreate.Pipeline.Slug, plan.ProviderSettings, p.client, timeouts)
@@ -657,7 +657,7 @@ func (p *pipelineResource) Update(ctx context.Context, req resource.UpdateReques
 		Tags:                                 getTagsFromSchema(&plan),
 	}
 
-	timeouts, diags := p.client.timeouts.Read(ctx, DefaultTimeout)
+	timeouts, diags := p.client.timeouts.Update(ctx, DefaultTimeout)
 	resp.Diagnostics.Append(diags...)
 	if resp.Diagnostics.HasError() {
 		return
@@ -728,7 +728,6 @@ func setPipelineModel(model *pipelineResourceModel, data pipelineResponse) {
 	model.SkipIntermediateBuildsBranchFilter = types.StringValue(data.GetSkipIntermediateBuildsBranchFilter())
 	model.Slug = types.StringValue(data.GetSlug())
 	model.Steps = types.StringValue(data.GetSteps().Yaml)
-	model.WebhookUrl = types.StringValue(data.GetWebhookURL())
 
 	tags := make([]types.String, len(data.GetTags()))
 	for i, tag := range data.GetTags() {