+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- SUP-1539: Limit the setting of Webhook url on Create only [[PR #450](https://github.com/buildkite/terraform-provider-buildkite/pull/450)]
+
 ## [v1.1.1](https://github.com/buildkite/terraform-provider-buildkite/compare/v1.1.1...v1.1.1)
 
 - Fix a typo in the signed pipelines data source [[PR #438](https://github.com/buildkite/terraform-provider-buildkite/pull/438)]